Moved much code into the library
