+++ .idea/compiler.xml
@@ -20,4 +20,3 @@
     </annotationProcessing>
   </component>
 </project>
-
