+++ BiscuitCaseApp/BiscuitCaseApp.iml
@@ -101,9 +101,9 @@
     <orderEntry type="library" exported="" scope="TEST" name="hamcrest-core-1.1" level="project" />
     <orderEntry type="library" exported="" scope="TEST" name="testing-support-lib-0.1" level="project" />
     <orderEntry type="library" exported="" scope="TEST" name="espresso-idling-resource-2.0" level="project" />
+    <orderEntry type="library" exported="" name="cardview-v7-22.1.0" level="project" />
     <orderEntry type="library" exported="" name="jsr305-2.0.1" level="project" />
-    <orderEntry type="library" exported="" name="cardview-v7-22.0.0" level="project" />
     <orderEntry type="library" exported="" name="fab-1.0.5" level="project" />
-    <orderEntry type="module" module-name="BiscuitCaseLibrary" />
+    <orderEntry type="module" module-name="BiscuitCaseLibrary" exported="" />
   </component>
 </module>
