+++ BiscuitCaseApp/build.gradle
@@ -20,13 +20,13 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile fileTree(include: ['*.jar'], dir: 'libs')
     compile 'com.android.support:appcompat-v7:22.1.0'
     compile 'com.android.support:recyclerview-v7:22.1.0'
     compile 'com.android.support:cardview-v7:22.0.0'
     compile 'com.github.shell-software:fab:1.0.5'
     compile 'com.facebook.stetho:stetho:1.0.1'
-
     androidTestCompile 'com.android.support.test.espresso:espresso-core:2.0'
     androidTestCompile 'com.android.support.test:testing-support-lib:0.1'
+    compile project(':BiscuitCaseLibrary')
 }
