--- BiscuitCaseApp/src/main/res/values/strings.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="app_name">HappyBurger</string>
-    <string name="hello_world">Hello world!</string>
-    <string name="action_settings">Settings</string>
-    <string name="totalLabelText">Total</string>
-    <string name="totalPlaceholder">"$0.00"</string>
-    <string name="title_activity_checkout">Checkout</string>
-
-<!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-
-</resources>
