+++ BiscuitCaseLibrary/BiscuitCaseLibrary.iml
@@ -87,7 +87,10 @@
     <orderEntry type="jdk" jdkName="Android API 22 Platform" jdkType="Android SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" exported="" name="appcompat-v7-22.1.0" level="project" />
+    <orderEntry type="library" exported="" name="recyclerview-v7-22.1.0" level="project" />
     <orderEntry type="library" exported="" name="support-annotations-22.1.0" level="project" />
     <orderEntry type="library" exported="" name="support-v4-22.1.0" level="project" />
+    <orderEntry type="library" exported="" name="cardview-v7-22.1.0" level="project" />
+    <orderEntry type="library" exported="" name="fab-1.0.5" level="project" />
   </component>
 </module>
