+++ BiscuitCaseLibrary/build.gradle
@@ -19,6 +19,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile fileTree(include: ['*.jar'], dir: 'libs')
     compile 'com.android.support:appcompat-v7:22.1.0'
+    compile 'com.android.support:cardview-v7:22.1.0'
+    compile 'com.github.shell-software:fab:1.0.5'
+    compile 'com.android.support:recyclerview-v7:22.1.0'
 }
