+++ BiscuitCaseLibrary/src/main/java/me/noahandrews/biscuitcaselibrary/Constants.java
@@ -1,0 +1,7 @@
+package me.noahandrews.biscuitcaselibrary;
+
+public class Constants {
+    public static final String WARNING_TAG = "BiscuitCaseWARNING";
+    public static final String DEBUG_TAG = "BiscuitCaseDEBUG";
+
+}
