+++ BiscuitCaseLibrary/src/main/res/values/dimens.xml
