+++ BiscuitCaseManagement/BiscuitCaseManagement.iml
@@ -87,5 +87,6 @@
     <orderEntry type="library" exported="" name="appcompat-v7-22.1.0" level="project" />
     <orderEntry type="library" exported="" name="support-annotations-22.1.0" level="project" />
     <orderEntry type="library" exported="" name="support-v4-22.1.0" level="project" />
+    <orderEntry type="module" module-name="BiscuitCaseLibrary" exported="" />
   </component>
 </module>
