+++ BiscuitCaseManagement/build.gradle
@@ -37,6 +37,7 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile fileTree(include: ['*.jar'], dir: 'libs')
     compile 'com.android.support:appcompat-v7:22.1.0'
+    compile project(':BiscuitCaseLibrary')
 }
