+++ BiscuitCaseManagement/src/main/AndroidManifest.xml
@@ -1,15 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="me.noahandrews.biscuitcasemanagement.app" >
+          package="me.noahandrews.biscuitcasemanagement.app"
+          android:versionCode="1"
+          android:versionName="0.1">
 
     <application
+        android:name=".MyApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
+        android:theme="@style/AppTheme">
         <activity
             android:name=".CategoryActivity"
-            android:label="@string/app_name" >
+            android:label="@string/app_name">
         </activity>
     </application>
